Move rgp module back into extensions/rgp.rs
