+++ epp-client/src/lib.rs
@@ -135,25 +135,7 @@ pub mod domain {
 pub mod extensions {
     pub mod consolidate;
     pub mod namestore;
-    pub mod rgp {
-        use serde::{Deserialize, Serialize};
-
-        pub mod report;
-        pub mod request;
-
-        pub const XMLNS: &str = "urn:ietf:params:xml:ns:rgp-1.0";
-
-        #[derive(Debug, Deserialize, Serialize)]
-        pub struct Update<T> {
-            #[serde(
-                rename = "rgp:update",
-                alias = "update",
-                alias = "upData",
-                alias = "infData"
-            )]
-            pub data: T,
-        }
-    }
+    pub mod rgp;
 }
 
 pub mod host {
